requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
ConfigArgParse==1.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.3
drf-yasg==1.20.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
greenlet==1.0.0
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
locust==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
msgpack==1.0.2
packaging==20.9
psutil==5.8.0
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
pywin32==300
pyzmq==22.0.3
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.3.0
